feat: Refactor language setup scripts to dynamically set primary and secondary languages, input methods, and time zones based on secure and inherited variables

+++ language-setup/language-setup-part1.ps1
@@ -11,7 +11,31 @@
 #>
 
 begin {
-    $primaryLanguage = 'en-GB'
+    $primaryLanguage = if (-not [string]::IsNullOrEmpty($SecureVars.primaryLanguage)) {
+        $SecureVars.primaryLanguage
+    }
+    else {
+        $InheritedVars.primaryLanguage
+    }
+
+    $secondaryLanguage = if (-not [string]::IsNullOrEmpty($SecureVars.secondaryLanguage)) {
+        $SecureVars.secondaryLanguage
+    }
+    else {
+        $InheritedVars.secondaryLanguage
+    }
+
+    $additionalLanguages = if (-not [string]::IsNullOrEmpty($SecureVars.additionalLanguages)) {
+        $SecureVars.additionalLanguages
+    }
+
+
+    [array]$languages = $primaryLanguage, $secondaryLanguage
+    if (!([string]::IsNullOrEmpty($additionalLanguages))) {
+        $languages += $additionalLanguages.Split(';')
+    }
+
+
     # Get OS Name
     $osName = (Get-ComputerInfo).OsName
     $os = if ($osName -match "Server \d+") {
@@ -25,15 +49,15 @@ begin {
     else {
         $osName
     }
-    $gitlab_root = "https://gitlab.com/ondemand-engineering"
-    $repo_root = "$gitlab_root/windows/windows-language-setup/-/raw/main/language_packs/$os"
-    $languagePackUri = "$repo_root/Microsoft-Windows-$type-Language-Pack_x64_$($primaryLanguage.toLower()).cab"
+    $gitlab_root = "https://raw.githubusercontent.com/ondemand-engineering"
+    $repo_root = "$gitlab_root/windows-language-setup/refs/heads/main/language_packs/$os"
 
     # Start powershell logging
     $SaveVerbosePreference = $VerbosePreference
     $VerbosePreference = 'continue'
     $VMTime = Get-Date
     $LogTime = $VMTime.ToUniversalTime()
+
 
     # Create the directory if it doesn't exist
     if (!(Test-Path -Path "$env:SYSTEMROOT\Temp\NerdioManagerLogs\ScriptedActions\languageSetup")) {
@@ -47,49 +71,67 @@ begin {
 }
 
 process {
-    # Language Pack Download
-    Write-Host "Downloading language pack"
-    Start-BitsTransfer -Source $languagePackUri -Destination "$env:SYSTEMROOT\Temp\$(Split-Path $languagePackUri -Leaf)"
-    Write-Host "Language pack downloaded"
-    $languagePack = Get-Item -Path "$env:SYSTEMROOT\Temp\$(Split-Path $languagePackUri -Leaf)"
-    Unblock-File -Path $languagePack.FullName -ErrorAction SilentlyContinue
+    # Disable Language Pack Cleanup
+    Disable-ScheduledTask -TaskPath "\Microsoft\Windows\AppxDeploymentClient\" -TaskName "Pre-staged app cleanup"
+    Disable-ScheduledTask -TaskPath "\Microsoft\Windows\MUI\" -TaskName "LPRemove"
+    Disable-ScheduledTask -TaskPath "\Microsoft\Windows\LanguageComponentsInstaller" -TaskName "Uninstallation"
+    reg add "HKEY_LOCAL_MACHINE\SOFTWARE\Policies\Microsoft\Control Panel\International" /v "BlockCleanupOfUnusedPreinstalledLangPacks" /t REG_DWORD /d 1 /f
 
-    # Install Language Pack
-    Write-Host "Installing language pack"
-    Add-WindowsPackage -Online -PackagePath $languagePack.FullName -NoRestart
-    Write-Host "Language pack installed"
+    foreach ($lang in ($languages | Where-Object { $_ -ne 'en-US' })) {
+        if (!(Get-WindowsPackage -Online | Where-Object { $_.ReleaseType -eq "LanguagePack" -and $_.PackageName -like "*LanguagePack*$lang*" })) {
 
-    # Remove Language Pack
-    $LanguagePack | Remove-Item -Force
-    Write-Host "Removed language pack cab file"
+            $languagePackUri = "$repo_root/Microsoft-Windows-$type-Language-Pack_x64_$($lang.toLower()).cab"
 
-    $capabilities = @(
-        "Microsoft-Windows-LanguageFeatures-Basic-$($primaryLanguage.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
-        "Microsoft-Windows-LanguageFeatures-Handwriting-$($primaryLanguage.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
-        "Microsoft-Windows-LanguageFeatures-OCR-$($primaryLanguage.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
-        "Microsoft-Windows-LanguageFeatures-Speech-$($primaryLanguage.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
-        "Microsoft-Windows-LanguageFeatures-TextToSpeech-$($primaryLanguage.toLower())-Package~31bf3856ad364e35~amd64~~.cab"
-    )
-    # Windows Capability Download
+            # Language Pack Download
+            Write-Host "Downloading language pack"
+            Start-BitsTransfer -Source $languagePackUri -Destination "$env:SYSTEMROOT\Temp\$(Split-Path $languagePackUri -Leaf)"
+            Write-Host "Language pack downloaded"
+            $languagePack = Get-Item -Path "$env:SYSTEMROOT\Temp\$(Split-Path $languagePackUri -Leaf)"
+            Unblock-File -Path $languagePack.FullName -ErrorAction SilentlyContinue
 
-    # Windows Capability Install
+            # Install Language Pack
+            Write-Host "Installing language pack"
+            Add-WindowsPackage -Online -PackagePath $languagePack.FullName -NoRestart
+            Write-Host "Language pack installed"
 
-    foreach ($Capability in $Capabilities) {
-        $capabilityUri = "$repo_root/$capability"
+            # Remove Language Pack
+            $LanguagePack | Remove-Item -Force
+            Write-Host "Removed language pack cab file"
+        }
 
-        Write-Host "Downloading $($Capability.Name)"
-        Start-BitsTransfer -Source $capabilityUri -Destination "$env:SYSTEMROOT\Temp\$(Split-Path $capabilityUri -Leaf)"
-        Write-Host "$($Capability.Name) downloaded"
-        $file = Get-Item -Path "$env:SYSTEMROOT\Temp\$(Split-Path $capabilityUri -Leaf)"
-        Unblock-File -Path $file.FullName -ErrorAction SilentlyContinue
+        if (($os -ne "server_2016") -or ($os -ne "server_2019")) {
 
-        Write-Host "Installing $($Capability.Name)"
-        Add-WindowsPackage -Online -PackagePath $file.FullName -NoRestart
-        Write-Host "$($Capability.Name) Installed"
+            $capabilities = @(
+                "Microsoft-Windows-LanguageFeatures-Basic-$($lang.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
+                "Microsoft-Windows-LanguageFeatures-Handwriting-$($lang.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
+                "Microsoft-Windows-LanguageFeatures-OCR-$($lang.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
+                "Microsoft-Windows-LanguageFeatures-Speech-$($lang.toLower())-Package~31bf3856ad364e35~amd64~~.cab",
+                "Microsoft-Windows-LanguageFeatures-TextToSpeech-$($lang.toLower())-Package~31bf3856ad364e35~amd64~~.cab"
+            )
+
+            foreach ($Capability in $Capabilities) {
+                if ((Get-WindowsCapability -Online | Where-Object { $_.Name -match "$lang" -and $_.Name -match $capability.Split("-")[3] }).State -ne "Installed") {
+
+                    $capabilityUri = "$repo_root/$capability"
+
+                    # Windows Capability Download
+                    Write-Host "Downloading $($Capability.Name)"
+                    Start-BitsTransfer -Source $capabilityUri -Destination "$env:SYSTEMROOT\Temp\$(Split-Path $capabilityUri -Leaf)"
+                    Write-Host "$($Capability.Name) downloaded"
+                    $file = Get-Item -Path "$env:SYSTEMROOT\Temp\$(Split-Path $capabilityUri -Leaf)"
+                    Unblock-File -Path $file.FullName -ErrorAction SilentlyContinue
+
+                    # Windows Capability Install
+                    Write-Host "Installing $($Capability.Name)"
+                    Add-WindowsPackage -Online -PackagePath $file.FullName -NoRestart
+                    Write-Host "$($Capability.Name) Installed"
+                }
+            }
+        }
     }
 
     # Set System Language
-    Set-WinSystemLocale -SystemLocale $primaryLanguage
+    Set-WinSystemLocale -SystemLocale $lang
     Write-Host "Set system locale"
 }
